+++ include/distortos/devices/communication/SpiMasterProxy.hpp
@@ -2,7 +2,7 @@
  * \file
  * \brief SpiMasterProxy class header
  *
- * \author Copyright (C) 2018 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ * \author Copyright (C) 2018-2019 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
  *
  * \par License
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
@@ -25,7 +25,7 @@ namespace devices
 {
 
 class SpiDevice;
-class SpiDeviceProxy;
+class SpiDeviceHandle;
 class SpiMaster;
 
 /**
@@ -48,10 +48,10 @@ public:
 	 *
 	 * \warning This function must not be called from interrupt context!
 	 *
-	 * \param [in] spiDeviceProxy is a reference to SpiDeviceProxy associated with this proxy
+	 * \param [in] spiDeviceHandle is a reference to SpiDeviceHandle associated with this proxy
 	 */
 
-	explicit SpiMasterProxy(const SpiDeviceProxy& spiDeviceProxy);
+	explicit SpiMasterProxy(const SpiDeviceHandle& spiDeviceHandle);
 
 	/**
 	 * \brief SpiMasterProxy's destructor
@@ -140,8 +140,8 @@ private:
 	/// range of transfers that are part of currently handled transaction
 	SpiMasterTransfersRange transfersRange_;
 
-	/// reference to SpiDeviceProxy associated with this proxy
-	const SpiDeviceProxy& spiDeviceProxy_;
+	/// reference to SpiDeviceHandle associated with this proxy
+	const SpiDeviceHandle& spiDeviceHandle_;
 
 	/// error codes detected in transferCompleteEvent()
 	volatile int ret_;